Move gblur tests to python/test/extra/ffmpeg_externaltest.py.
