+++ python/test/extra/noref_feature_extractor_extratest.py
@@ -63,29 +63,6 @@ class NorefFeatureExtractorTest(unittest.TestCase):
         self.assertAlmostEqual(results[0]['Moment_noref_feature_2nd_score'], 4888.7623296039092, places=4)
         self.assertAlmostEqual(results[0]['Moment_noref_feature_var_score'], 1005.5413716918079, places=4)
 
-    def test_noref_moment_fextractor_with_noref_asset_notyuv_gaussianblur(self):
-
-        dis_path = VmafConfig.test_resource_path("mp4", "Seeking_10_288_375.mp4")
-        asset = NorefAsset(dataset="test", content_id=0, asset_id=0,
-                      workdir_root=VmafConfig.workdir_path(),
-                      dis_path=dis_path,
-                      asset_dict={'yuv_type': 'notyuv',
-                                  'quality_width': 720, 'quality_height': 480,
-                                  'gblur_cmd': 'sigma=0.01:steps=1',
-                                  })
-
-        self.fextractor = MomentNorefFeatureExtractor(
-            [asset],
-            None, fifo_mode=True,
-            result_store=None
-        )
-        self.fextractor.run()
-
-        results = self.fextractor.results
-
-        self.assertAlmostEqual(results[0]['Moment_noref_feature_1st_score'], 63.273978452932084, places=4)
-        self.assertAlmostEqual(results[0]['Moment_noref_feature_2nd_score'], 5124.572291840278, places=4)
-        self.assertAlmostEqual(results[0]['Moment_noref_feature_var_score'], 1111.9962740092349, places=4)
 
 if __name__ == '__main__':
     unittest.main(verbosity=2)
